requirements.txt
ccxt>=3.0.0
pandas>=1.5.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.28.0